+++ compiler-plugin/src/main/kotlin/tech/mappie/fir/analysis/CompileTimeDslReceiverChecker.kt
@@ -1,8 +1,6 @@
 package tech.mappie.fir.analysis
 
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
-import org.jetbrains.kotlin.diagnostics.SourceElementPositioningStrategies.WHOLE_ELEMENT
-import org.jetbrains.kotlin.diagnostics.error1
 import org.jetbrains.kotlin.diagnostics.reportOn
 import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
@@ -10,12 +8,15 @@ import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirFunctionCallChec
 import org.jetbrains.kotlin.fir.expressions.FirFunctionCall
 import org.jetbrains.kotlin.fir.types.classId
 import org.jetbrains.kotlin.fir.types.resolvedType
-import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.psi.KtElement
+import tech.mappie.fir.analysis.MappieErrors.COMPILE_TIME_EXTENSION_RECEIVER
+import tech.mappie.fir.analysis.MappieErrors.COMPILE_TIME_RECEIVER
 import tech.mappie.util.ALL_MAPPING_FUNCTIONS
+import tech.mappie.util.CLASS_ID_MULTIPLE_OBJECT_MAPPING_CONSTRUCTOR
 import tech.mappie.util.CLASS_ID_OBJECT_MAPPING_CONSTRUCTOR
 
 class CompileTimeDslReceiverChecker : FirFunctionCallChecker(MppCheckerKind.Common) {
+
+    private val targets = listOf(CLASS_ID_OBJECT_MAPPING_CONSTRUCTOR, CLASS_ID_MULTIPLE_OBJECT_MAPPING_CONSTRUCTOR)
 
     context(context: CheckerContext, reporter: DiagnosticReporter)
     override fun check(expression: FirFunctionCall) {
@@ -24,37 +25,12 @@ class CompileTimeDslReceiverChecker : FirFunctionCallChecker(MppCheckerKind.Comm
             return
         }
 
-        if (expression.dispatchReceiver?.resolvedType?.classId == CLASS_ID_OBJECT_MAPPING_CONSTRUCTOR) {
-            reporter.reportOn(
-                expression.source,
-                COMPILE_TIME_RECEIVER,
-                buildString {
-                    append("The function $name was called on the mapping dsl which does not exist after compilation")
-                    specify(name)
-                },
-            )
+        if (expression.dispatchReceiver?.resolvedType?.classId in targets) {
+            reporter.reportOn(expression.source, COMPILE_TIME_RECEIVER, name)
         }
 
-        if (expression.extensionReceiver?.resolvedType?.classId == CLASS_ID_OBJECT_MAPPING_CONSTRUCTOR) {
-            reporter.reportOn(
-                expression.source,
-                COMPILE_TIME_RECEIVER,
-                buildString {
-                    append("The function $name was called as an extension method on the mapping dsl which does not exist after compilation")
-                    specify(name)
-                }
-            )
+        if (expression.extensionReceiver?.resolvedType?.classId in targets) {
+            reporter.reportOn(expression.source, COMPILE_TIME_EXTENSION_RECEIVER, name)
         }
-    }
-
-    private fun StringBuilder.specify(name: Name) =
-        if (name == Name.identifier("run")) {
-            append(". Did you mean to use kotlin.run?")
-        } else {
-            this
-        }
-
-    companion object {
-        private val COMPILE_TIME_RECEIVER by error1<KtElement, String>(WHOLE_ELEMENT)
     }
 }