+++ compiler-plugin/src/main/kotlin/tech/mappie/ir/resolving/classes/ClassResolver.kt
@@ -1,10 +1,12 @@
 package tech.mappie.ir.resolving.classes
 
 import org.jetbrains.kotlin.ir.declarations.IrFunction
+import org.jetbrains.kotlin.ir.expressions.IrFunctionExpression
 import tech.mappie.ir.resolving.*
 import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.types.getClass
 import org.jetbrains.kotlin.ir.util.constructors
+import org.jetbrains.kotlin.ir.util.statements
 import org.jetbrains.kotlin.name.Name
 import tech.mappie.ir.resolving.classes.targets.MappieTargetsCollector
 
@@ -19,7 +21,13 @@ class ClassResolver(
             ClassMappingRequestBuilder(constructor, context)
                 .targets(MappieTargetsCollector(target, function, constructor).collect())
                 .sources(sources)
-                .also { function?.body?.accept(ExplicitClassMappingCollector(context), it) }
+                .apply {
+                    val mapping = findMappingStatements(function?.body).singleOrNull()?.arguments?.getOrNull(1) as? IrFunctionExpression
+                    mapping?.function?.body?.statements?.forEach { statement ->
+                        statement.accept(ClassMappingStatementCollector(context), Unit)
+                            ?.let { explicit(it) }
+                    }
+                }
                 .construct()
         }.toList()
 }