+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/enums/ObjectMappieInsteadOfEnumMappieTest.kt
@@ -37,11 +37,11 @@ class ObjectMappieInsteadOfEnumMappieTest : MappieTestCase() {
 
                 class Mapper : ObjectMappie<InputObject, Output>() {
                     override fun map(from: InputObject) = 
-                    if (from.boolean) Output.TRUE else Output.FALSE
+                        if (from.boolean) Output.TRUE else Output.FALSE
                 }
                 """
             )
-        } satisfies  {
+        } satisfies {
             isOk()
         }
     }