+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/objects/ConstructorParameterNotAFieldTest.kt
@@ -32,7 +32,7 @@ class ConstructorParameterNotAFieldTest : MappieTestCase() {
             )
         } satisfies {
             isCompilationError()
-            hasErrorMessage(6, "Identifier fake does not occur as as setter or as a parameter in constructor")
+            hasErrorMessage(6, "Identifier 'fake' does not occur as as setter or as a parameter in constructor")
         }
     }