+++ website/src/changelog.md
@@ -3,6 +3,19 @@ title: "Changelog"
 layout: "layouts/changelog.html"
 changelog:
   - date: "tbd"
+    title: "v2.2.20-1.6.0"
+    items:
+      - "Allow `mapping` to be used anywhere (at most once) in the body of the map function."
+      - "Fixed check of runtime usage of of mapping DSL for multiple object mappie."
+      - "Minor performance enhancements - by @ErwinOlie."
+      - "Fix for missing reporting when both a mapper and its generated mapper have issues - by @ErwinOlie."
+      - "Improved reporting of type parameters - by @ErwinOlie."
+      - "`AbstractList` and `AbstractMutableList` are now supported as source type for mapping to `List` - by @ErwinOlie."
+  - date: "2025-09-10"
+    title: "v2.2.20-1.5.0"
+    items:
+      - "Bumped Kotlin to version 2.2.20."
+  - date: "2025-09-10"
     title: "v2.2.10-1.5.0"
     items:
       - "Added support for kotlinx-datetime to JVM mappers, which can be imported via a separate dependency."